+++ src/App.jsx
@@ -1,49 +1,15 @@
 import React, { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
 import './App.css';
 import Header from './components/Header';
 import Form from './components/Form';
 import TodoList from './components/TodoList';
 
 function App() {
-  const id = uuidv4();
-  // 값 변경 + 리렌더링을 위한 State 선언 + 로컬스토리지
   const [todos, setTodos] = useState(() => JSON.parse(localStorage.getItem('todos')) || []);
   localStorage.setItem('todos', JSON.stringify(todos));
   const [doneTodos, setDoneTodo] = useState(() => JSON.parse(localStorage.getItem('doneTodos')) || []);
   localStorage.setItem('doneTodos', JSON.stringify(doneTodos));
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
 
-  const titleChangeHandler = (e) => {
-    setTitle(e.target.value);
-  };
-
-  const contentChangeHandler = (e) => {
-    setContent(e.target.value);
-  };
-
-  const clickAddBtnHandler = (e) => {
-    e.preventDefault();
-    // 불필요한 새로고침 동작 막기
-
-    if (!title) return alert('제목을 입력해주세요');
-    if (!content) return alert('내용을 입력해주세요');
-    // 빈 값 검사
-
-    const newTodos = {
-      id,
-      title,
-      content,
-    };
-    // 좌, 우 키값이 동일할 경우 생략 가능
-
-    setTodos([...todos, newTodos]);
-    setTitle('');
-    setContent('');
-    // 기존 배열(todos)을 풀고 newTodos를 더해 새로운 배열로 변환
-    // 제목, 내용 인풋값 초기화
-  };
   // '삭제' 버튼 클릭 이벤트
   const clickDelBtnHandler = (id) => {
     const newWorkingTodos = todos.filter((todo) => todo.id !== id);
@@ -70,7 +36,7 @@ function App() {
   return (
     <div className="layout">
       <Header />
-      <Form title={title} content={content} titleChangeHandler={titleChangeHandler} contentChangeHandler={contentChangeHandler} clickAddBtnHandler={clickAddBtnHandler} />
+      <Form todos={todos} setTodos={setTodos} />
       <div className="list-container">
         <h2 className="list-title">Working...🔥</h2>
         <div className="todo-wrap">
